refactor: remove activity lifecycle listener
