+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/session/lifecycle/ActivityLifecycleTracker.kt
@@ -3,28 +3,18 @@ package io.embrace.android.embracesdk.internal.session.lifecycle
 import android.app.Activity
 import android.app.Application
 import android.os.Bundle
-import io.embrace.android.embracesdk.internal.logging.EmbLogger
-import io.embrace.android.embracesdk.internal.logging.InternalErrorType
 import java.lang.ref.WeakReference
-import java.util.concurrent.CopyOnWriteArrayList
 
 /**
  * Tracks activity lifecycle events.
  */
 class ActivityLifecycleTracker(
     private val application: Application,
-    private val logger: EmbLogger,
 ) : ActivityTracker {
 
     init {
         application.registerActivityLifecycleCallbacks(this)
     }
-
-    /**
-     * List of listeners that subscribe to activity events.
-     */
-    val activityListeners: CopyOnWriteArrayList<ActivityLifecycleListener> =
-        CopyOnWriteArrayList<ActivityLifecycleListener>()
 
     /**
      * The currently active activity.
@@ -57,53 +47,25 @@ class ActivityLifecycleTracker(
 
     override fun onActivityCreated(activity: Activity, bundle: Bundle?) {
         updateStateWithActivity(activity)
-        activityListeners.toList().forEach { listener: ActivityLifecycleListener ->
-            try {
-                listener.onActivityCreated(activity, bundle)
-            } catch (ex: Exception) {
-                logger.trackInternalError(InternalErrorType.ACTIVITY_LISTENER_FAIL, ex)
-            }
-        }
     }
 
     override fun onActivityStarted(activity: Activity) {
         updateStateWithActivity(activity)
-        activityListeners.toList().forEach { listener: ActivityLifecycleListener ->
-            try {
-                listener.onActivityStarted(activity)
-            } catch (ex: Exception) {
-                logger.trackInternalError(InternalErrorType.ACTIVITY_LISTENER_FAIL, ex)
-            }
-        }
     }
 
     override fun onActivityResumed(activity: Activity) {}
     override fun onActivityPaused(activity: Activity) {}
 
     override fun onActivityStopped(activity: Activity) {
-        activityListeners.toList().forEach { listener: ActivityLifecycleListener ->
-            try {
-                listener.onActivityStopped(activity)
-            } catch (ex: Exception) {
-                logger.trackInternalError(InternalErrorType.ACTIVITY_LISTENER_FAIL, ex)
-            }
-        }
     }
 
     override fun onActivitySaveInstanceState(activity: Activity, bundle: Bundle) {}
 
     override fun onActivityDestroyed(activity: Activity) {}
 
-    override fun addListener(listener: ActivityLifecycleListener) {
-        if (!activityListeners.contains(listener)) {
-            activityListeners.addIfAbsent(listener)
-        }
-    }
-
     override fun close() {
         runCatching {
             application.unregisterActivityLifecycleCallbacks(this)
-            activityListeners.clear()
         }
     }
 }
