--- embrace-android-instrumentation-api/src/main/kotlin/io/embrace/android/embracesdk/internal/session/lifecycle/ActivityLifecycleListener.kt
@@ -1,25 +1,0 @@
-package io.embrace.android.embracesdk.internal.session.lifecycle
-
-import android.app.Activity
-import android.app.Application.ActivityLifecycleCallbacks
-import android.os.Bundle
-
-/**
- * Implementation of [ActivityLifecycleCallbacks] with no-op defaults
- */
-interface ActivityLifecycleListener : ActivityLifecycleCallbacks {
-
-    override fun onActivityCreated(activity: Activity, bundle: Bundle?) {}
-
-    override fun onActivityStarted(activity: Activity) {}
-
-    override fun onActivityResumed(activity: Activity) {}
-
-    override fun onActivityPaused(activity: Activity) {}
-
-    override fun onActivityStopped(activity: Activity) {}
-
-    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {}
-
-    override fun onActivityDestroyed(activity: Activity) {}
-}
